Refactoring out the user location manager
