+++ d2d-territory-management-service/MapSearchView.swift
@@ -804,32 +804,3 @@ struct MapSearchView: View {
 extension MapSearchView {
     struct KnockStepperState: Identifiable, Equatable { let id = UUID(); var ctx: KnockContext }
 }
-
-final class UserLocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
-    private let manager = CLLocationManager()
-
-    @Published var heading: CLHeading?
-    @Published var location: CLLocation?
-
-    override init() {
-        super.init()
-        manager.delegate = self
-        manager.desiredAccuracy = kCLLocationAccuracyBest
-        manager.headingFilter = 1
-        manager.requestWhenInUseAuthorization()
-        manager.startUpdatingLocation()
-        manager.startUpdatingHeading()
-    }
-
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        location = locations.last
-    }
-
-    func locationManager(_ manager: CLLocationManager, didUpdateHeading newHeading: CLHeading) {
-        heading = newHeading
-    }
-
-    func locationManagerShouldDisplayHeadingCalibration(_ manager: CLLocationManager) -> Bool {
-        true
-    }
-}
