+++ d2d-territory-management-service/controllers/UserLocationManager.swift
@@ -1,0 +1,52 @@
+//
+//  UserLocationManager.swift
+//  d2d-studio
+//
+//  Created by Emin Okic on 12/20/25.
+//
+
+import Foundation
+import CoreLocation
+import Combine
+
+final class UserLocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+
+    private let manager = CLLocationManager()
+
+    @Published var heading: CLHeading?
+    @Published var location: CLLocation?
+
+    override init() {
+        super.init()
+
+        manager.delegate = self
+        manager.desiredAccuracy = kCLLocationAccuracyBest
+        manager.headingFilter = 1
+
+        manager.requestWhenInUseAuthorization()
+        manager.startUpdatingLocation()
+        manager.startUpdatingHeading()
+    }
+
+    // MARK: - CLLocationManagerDelegate
+
+    func locationManager(
+        _ manager: CLLocationManager,
+        didUpdateLocations locations: [CLLocation]
+    ) {
+        location = locations.last
+    }
+
+    func locationManager(
+        _ manager: CLLocationManager,
+        didUpdateHeading newHeading: CLHeading
+    ) {
+        heading = newHeading
+    }
+
+    func locationManagerShouldDisplayHeadingCalibration(
+        _ manager: CLLocationManager
+    ) -> Bool {
+        true
+    }
+}
